management_console/backend/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.1
PyJWT==1.7.1
djangorestframework-simplejwt==4.4.0
phonenumbers==8.12.20
Pillow==6.2.2
psycopg2-binary==2.8.6
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
django-filter==2.4.0